feat: added decorator with object class definition
fix: added api endpoint
