+++ MANIFEST.in
@@ -18,5 +18,6 @@ include CHANGELOG.rst
 include CONTRIBUTING.rst
 include LICENSE
 include README.rst
+include CODE_OF_CONDUCT.md
 
 global-exclude *.py[cod] __pycache__/* *.so *.dylib
