+++ src/quads_lib/decorators.py
@@ -1,0 +1,10 @@
+def returns(resource_type):
+    """
+    Decorator to annotate methods with the resource type they return.
+    """
+
+    def decorator(func):
+        func._returns = resource_type
+        return func
+
+    return decorator
