+++ src/quads_lib/exceptions.py
@@ -1,0 +1,6 @@
+class APIServerException(Exception):
+    pass
+
+
+class APIBadRequest(Exception):
+    pass
