+++ src/quads_lib/quads.py
@@ -1,81 +1,11 @@
-from json import JSONDecodeError
 from pathlib import Path
 from typing import Optional
 from urllib import parse as url_parse
 from urllib.parse import urlencode
 from urllib.parse import urljoin
 
-from requests import Session
-from requests.adapters import HTTPAdapter
-from requests.adapters import Retry
-from requests.auth import HTTPBasicAuth
-
-
-class APIServerException(Exception):
-    pass
-
-
-class APIBadRequest(Exception):
-    pass
-
-
-class QuadsBase:
-    """
-    Base class for the Quads API
-    """
-
-    def __init__(self, username: str, password: str, base_url: str):
-        self.username = username
-        self.password = password
-        self.base_url = base_url
-        self.session = Session()
-        retries = Retry(total=5, backoff_factor=1, status_forcelist=[502, 503, 504])
-        self.session.mount("http://", HTTPAdapter(max_retries=retries))
-        self.auth = HTTPBasicAuth(self.username, self.password)
-        self.token = None
-        self.headers = {}
-
-    def __enter__(self):
-        self.login()
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.logout()
-        self.session.close()
-
-    def _make_request(self, method: str, endpoint: str, data: Optional[dict] = None) -> dict:
-        _response = self.session.request(
-            method,
-            urljoin(self.base_url, endpoint),
-            json=data,
-            verify=False,
-        )
-        if _response.status_code == 500:
-            raise APIServerException("Check the flask server logs")
-        if _response.status_code == 400:
-            try:
-                response_json = _response.json()
-            except JSONDecodeError as e:
-                raise APIBadRequest("Failed to parse response") from e
-            raise APIBadRequest(response_json.get("message"))
-        return _response.json()
-
-    # Base functions
-    def get(self, endpoint: str) -> dict:
-        _response = self._make_request("GET", endpoint)
-        return _response
-
-    def post(self, endpoint: str, data: Optional[dict] = None) -> dict:
-        _response = self._make_request("POST", endpoint, data)
-        return _response
-
-    def patch(self, endpoint: str, data: Optional[dict] = None) -> dict:
-        _response = self._make_request("PATCH", endpoint, data)
-        return _response
-
-    def delete(self, endpoint: str) -> dict:
-        _response = self._make_request("DELETE", endpoint)
-        return _response
+from quads_lib.base import QuadsBase
+from quads_lib.decorators import returns
 
 
 class QuadsApi(QuadsBase):
@@ -105,6 +35,7 @@ class QuadsApi(QuadsBase):
         return json_response
 
     # Hosts
+    @returns("List[Host]")
     def get_hosts(self) -> dict:
         json_response = self.get("hosts")
         return json_response
@@ -113,30 +44,36 @@ class QuadsApi(QuadsBase):
         json_response = self.get("hosts?group_by=model")
         return json_response
 
+    @returns("List[Host]")
     def filter_hosts(self, data: dict) -> dict:
         url_params = url_parse.urlencode(data)
         json_response = self.get(f"hosts?{url_params}")
         return json_response
 
+    @returns("List[Cloud]")
     def filter_clouds(self, data: dict) -> dict:
         url_params = url_parse.urlencode(data)
         json_response = self.get(f"clouds?{url_params}")
         return json_response
 
+    @returns("List[Assignment]")
     def filter_assignments(self, data: dict) -> dict:
         url_params = url_parse.urlencode(data)
         json_response = self.get(f"assignments?{url_params}")
         return json_response
 
+    @returns("Host")
     def get_host(self, hostname: str) -> dict:
         endpoint = Path("hosts") / hostname
         json_response = self.get(str(endpoint))
         return json_response
 
+    @returns("Host")
     def create_host(self, data: dict) -> dict:
         json_response = self.post("hosts", data)
         return json_response
 
+    @returns("Host")
     def update_host(self, hostname: str, data: dict) -> dict:
         endpoint = Path("hosts") / hostname
         json_response = self.patch(str(endpoint), data)
@@ -154,14 +91,17 @@ class QuadsApi(QuadsBase):
         return True if "true" in json_response else False
 
     # Clouds
+    @returns("List[Cloud]")
     def get_clouds(self) -> dict:
         json_response = self.get("clouds")
         return json_response
 
+    @returns("List[Cloud]")
     def get_free_clouds(self) -> dict:
         json_response = self.get("clouds/free/")
         return json_response
 
+    @returns("Cloud")
     def get_cloud(self, cloud_name: str) -> dict:
         json_response = self.get(f"clouds?name={cloud_name}")
         return json_response
@@ -175,9 +115,11 @@ class QuadsApi(QuadsBase):
         json_response = self.get(url)
         return json_response
 
+    @returns("Cloud")
     def create_cloud(self, data: dict) -> dict:
         return self.post("clouds", data)
 
+    @returns("Cloud")
     def update_cloud(self, cloud_name: str, data: dict) -> dict:
         endpoint = Path("clouds") / cloud_name
         json_response = self.patch(str(endpoint), data)
@@ -189,6 +131,7 @@ class QuadsApi(QuadsBase):
         return json_response
 
     # Schedules
+    @returns("List[Schedule]")
     def get_schedules(self, data: Optional[dict] = None) -> dict:
         if data is None:
             data = {}
@@ -199,6 +142,7 @@ class QuadsApi(QuadsBase):
         json_response = self.get(url)
         return json_response
 
+    @returns("List[Schedule]")
     def get_current_schedules(self, data: Optional[dict] = None) -> dict:
         if data is None:
             data = {}
@@ -210,11 +154,13 @@ class QuadsApi(QuadsBase):
         json_response = self.get(url)
         return json_response
 
+    @returns("Schedule")
     def get_schedule(self, schedule_id: int) -> dict:
         endpoint = Path("schedules") / str(schedule_id)
         json_response = self.get(str(endpoint))
         return json_response
 
+    @returns("List[Schedule]")
     def get_future_schedules(self, data: Optional[dict] = None) -> dict:
         if data is None:
             data = {}
@@ -226,6 +172,7 @@ class QuadsApi(QuadsBase):
         json_response = self.get(url)
         return json_response
 
+    @returns("Schedule")
     def update_schedule(self, schedule_id: int, data: dict) -> dict:
         endpoint = Path("schedules") / str(schedule_id)
         json_response = self.patch(str(endpoint), data)
@@ -236,41 +183,50 @@ class QuadsApi(QuadsBase):
         json_response = self.delete(str(endpoint))
         return json_response
 
+    @returns("Schedule")
     def create_schedule(self, data: dict) -> dict:
         return self.post("schedules", data)
 
     # Available
+    @returns("List[Host]")
     def get_available(self) -> dict:
         json_response = self.get("available")
         return json_response
 
+    @returns("List[Host]")
     def filter_available(self, data: dict) -> dict:
         json_response = self.get(f"available?{urlencode(data)}")
         return json_response
 
     # Assignments
+    @returns("Assignment")
     def create_assignment(self, data: dict) -> dict:
         return self.post("assignments", data)
 
+    @returns("Assignment")
     def create_self_assignment(self, data: dict) -> dict:
         endpoint = Path("assignments") / "self"
         return self.post(str(endpoint), data)
 
+    @returns("Assignment")
     def update_assignment(self, assignment_id: int, data: dict) -> dict:
         endpoint = Path("assignments") / str(assignment_id)
         json_response = self.patch(str(endpoint), data)
         return json_response
 
+    @returns("Notification")
     def update_notification(self, notification_id: int, data: dict) -> dict:
         endpoint = Path("notifications") / str(notification_id)
         json_response = self.patch(str(endpoint), data)
         return json_response
 
+    @returns("Assignment")
     def get_active_cloud_assignment(self, cloud_name: str) -> dict:
         endpoint = Path("assignments") / "active" / cloud_name
         json_response = self.get(str(endpoint))
         return json_response
 
+    @returns("List[Assignment]")
     def get_active_assignments(self) -> dict:
         json_response = self.get("assignments/active")
         return json_response
@@ -281,15 +237,18 @@ class QuadsApi(QuadsBase):
         return json_response
 
     # Interfaces
+    @returns("List[Interface]")
     def get_host_interface(self, hostname: str) -> dict:
         endpoint = Path("hosts") / hostname / "interfaces"
         json_response = self.get(str(endpoint))
         return json_response
 
+    @returns("List[Interface]")
     def get_interfaces(self) -> dict:
         json_response = self.get("interfaces")
         return json_response
 
+    @returns("Interface")
     def update_interface(self, hostname: str, data: dict) -> dict:
         endpoint = Path("interfaces") / hostname
         json_response = self.patch(str(endpoint), data)
@@ -300,12 +259,14 @@ class QuadsApi(QuadsBase):
         json_response = self.delete(str(endpoint))
         return json_response
 
+    @returns("Interface")
     def create_interface(self, hostname: str, data: dict) -> dict:
         endpoint = Path("interfaces") / hostname
         json_response = self.post(str(endpoint), data)
         return json_response
 
     # Memory
+    @returns("Memory")
     def create_memory(self, hostname: str, data: dict) -> dict:
         endpoint = Path("memory") / hostname
         json_response = self.post(str(endpoint), data)
@@ -317,11 +278,13 @@ class QuadsApi(QuadsBase):
         return json_response
 
     # Disks
+    @returns("Disk")
     def create_disk(self, hostname: str, data: dict) -> dict:
         endpoint = Path("disks") / hostname
         json_response = self.post(str(endpoint), data)
         return json_response
 
+    @returns("Disk")
     def update_disk(self, hostname: str, data: dict) -> dict:
         endpoint = Path("disks") / hostname
         json_response = self.patch(str(endpoint), data)
@@ -333,6 +296,7 @@ class QuadsApi(QuadsBase):
         return json_response
 
     # Processor
+    @returns("Processor")
     def create_processor(self, hostname: str, data: dict) -> dict:
         endpoint = Path("processors") / hostname
         json_response = self.post(str(endpoint), data)
@@ -344,25 +308,30 @@ class QuadsApi(QuadsBase):
         return json_response
 
     # Vlans
+    @returns("List[Vlan]")
     def get_vlans(self) -> dict:
         json_response = self.get("vlans")
         return json_response
 
+    @returns("Vlan")
     def get_vlan(self, vlan_id: int) -> dict:
         endpoint = Path("vlans") / str(vlan_id)
         json_response = self.get(str(endpoint))
         return json_response
 
+    @returns("List[Vlan]")
     def get_free_vlans(self) -> dict:
         endpoint = Path("vlans") / "free"
         json_response = self.get(str(endpoint))
         return json_response
 
+    @returns("Vlan")
     def update_vlan(self, vlan_id: int, data: dict) -> dict:
         endpoint = Path("vlans") / str(vlan_id)
         json_response = self.patch(str(endpoint), data)
         return json_response
 
+    @returns("Vlan")
     def create_vlan(self, data: dict) -> dict:
         return self.post("vlans", data)
